Add files via upload
chatgpt: "your task is to add a border or box to contain each resume section. ensure this border/box goes the entirety of the horizontal layout (meaning from left side of page to the right, past the text). the border/box should be filled with color and sent to the back of the page so that the text within can be read on top of it"